backend/requirements.txt
fastapi==0.111.1
uvicorn==0.25.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.7
python-dotenv==1.0.1
pydantic==2.2.1
imapclient==2.3.0
pyzmail36==1.0.5
email-validator==2.0.0
